graph: Move InputSchema validations over from Schema
These validations only makes sense for input schemas, not generic GraphQL
schemas
